+++ IWFY_VDP2020_UNITY/.idea/.idea.IWFY_VDP2020_UNITY/.idea/contentModel.xml
@@ -31,9 +31,9 @@
             <e p="Jump.cs" t="Include" />
             <e p="SimpleJump.cs" t="Include" />
             <e p="Vibration.cs" t="Include" />
-            <e p="WIP_IwfyCameraMouseMapping.cs" t="Include" />
             <e p="WIP_IwfyCameraPointer.cs" t="Include" />
             <e p="WIP_IwfyClickableObject.cs" t="Include" />
+            <e p="WIP_IwfyPCMapping.cs" t="Include" />
             <e p="WIP_IwfyReticleAnimator.cs" t="Include" />
           </e>
         </e>
