+++ IWFY_VDP2020_UNITY/.idea/.idea.IWFY_VDP2020_UNITY/.idea/workspace.xml
@@ -26,14 +26,15 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="8cad7203-29f4-48a7-abb2-d4e843b989f7" name="Default Changelist" comment="">
+      <change afterPath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/WIP_IwfyPCMapping.cs" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/WIP_IwfyPCMapping.cs.meta" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/.idea.IWFY_VDP2020_UNITY/.idea/contentModel.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/.idea.IWFY_VDP2020_UNITY/.idea/contentModel.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/.idea.IWFY_VDP2020_UNITY/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/.idea.IWFY_VDP2020_UNITY/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Assets/IWFY_Experiments/Materials/SimpleRed.mat" beforeDir="false" afterPath="$PROJECT_DIR$/Assets/IWFY_Experiments/Materials/SimpleRed.mat" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scenes/MyExperiments 1.unity" beforeDir="false" afterPath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scenes/MyExperiments 1.unity" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scenes/ReticlePointer.unity" beforeDir="false" afterPath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scenes/ReticlePointer.unity" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/SimpleJump.cs" beforeDir="false" afterPath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/SimpleJump.cs" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/Assets/Samples/Google Cardboard XR Plugin for Unity/1.3.0/Hello Cardboard/Scripts/CameraPointer.cs" beforeDir="false" afterPath="$PROJECT_DIR$/Assets/Samples/Google Cardboard XR Plugin for Unity/1.3.0/Hello Cardboard/Scripts/CameraPointer.cs" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/WIP_IwfyCameraMouseMapping.cs" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/WIP_IwfyCameraMouseMapping.cs.meta" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/ProjectSettings/EditorBuildSettings.asset" beforeDir="false" afterPath="$PROJECT_DIR$/ProjectSettings/EditorBuildSettings.asset" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/ProjectSettings/GraphicsSettings.asset" beforeDir="false" afterPath="$PROJECT_DIR$/ProjectSettings/GraphicsSettings.asset" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/ProjectSettings/ProjectSettings.asset" beforeDir="false" afterPath="$PROJECT_DIR$/ProjectSettings/ProjectSettings.asset" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -66,6 +67,8 @@
         <option value="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/IwfyCameraMouseMapping.cs" />
         <option value="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/SimpleJump.cs" />
         <option value="$PROJECT_DIR$/Assets/Samples/Google Cardboard XR Plugin for Unity/1.3.0/Hello Cardboard/Scripts/CameraPointer.cs" />
+        <option value="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/WIP_IwfyPCMapping.cs" />
+        <option value="$PROJECT_DIR$/Assets/IWFY_Experiments/Scripts/WIP_IwfyCameraPointer.cs" />
       </list>
     </option>
   </component>
@@ -121,6 +124,7 @@
       <option name="presentableId" value="Default" />
       <updated>1604837065891</updated>
       <workItem from="1604837078990" duration="3946000" />
+      <workItem from="1604866610669" duration="721000" />
     </task>
     <servers />
   </component>
