+++ IWFY_VDP2020_UNITY/Assets/IWFY_Experiments/Scenes/ReticlePointer.unity
@@ -182,10 +182,8 @@ GameObject:
   - component: {fileID: 273391863}
   - component: {fileID: 273391862}
   - component: {fileID: 273391865}
-  - component: {fileID: 273391867}
   - component: {fileID: 273391866}
   - component: {fileID: 273391868}
-  - component: {fileID: 273391869}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -294,26 +292,6 @@ MonoBehaviour:
   speedV: 2
   pitch: 0
   yaw: 0
---- !u!114 &273391867
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 273391861}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 113e4aa2cd3db44d1ac0cb7e9794964b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  globalGainDb: 0
-  occlusionMask:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  stereoSpeakerModeEnabled: 0
-  recorderSeamless: 0
-  recorderSourceTag: Untagged
-  recorderFoldout: 0
 --- !u!114 &273391868
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -326,26 +304,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 908d8f59d3b47f88a8e0404d365d496a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &273391869
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 273391861}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d8ee29649a9ab3e49a0f9919d1fc8fa8, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  makeAllButtonsClickable: 1
-  raycastEveryUpdate: 1
-  interactablesLayers:
-    serializedVersion: 2
-    m_Bits: 4294967295
-  OnTriggerPressed:
-    m_PersistentCalls:
-      m_Calls: []
 --- !u!1 &553669769
 GameObject:
   m_ObjectHideFlags: 0
