--- IWFY_VDP2020_UNITY/Assets/IWFY_Experiments/Scripts/WIP_IwfyCameraMouseMapping.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 436db5202d9103c1a9cfc60e7ba653bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
