+++ IWFY_VDP2020_UNITY/Assets/IWFY_Experiments/Scripts/WIP_IwfyPCMapping.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 436db5202d9103c1a9cfc60e7ba653bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
