+++ IWFY_VDP2020_UNITY/ProjectSettings/EditorBuildSettings.asset
@@ -8,10 +8,10 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/IWFY_Experiments/Scenes/MyExperiments 1.unity
     guid: d7c29633a1159021daabd6f6a1ac2d51
-  - enabled: 0
+  - enabled: 1
     path: Assets/IWFY_Experiments/Scenes/ReticlePointer.unity
     guid: 1a9801b138b4eb963977998f3a1f8b34
-  - enabled: 1
+  - enabled: 0
     path: Assets/Samples/Google Cardboard XR Plugin for Unity/1.3.0/Hello Cardboard/Scenes/HelloCardboard.unity
     guid: 145a2a45d7bc746b19cb68410368bf96
   m_configObjects:
